Re-land part of r55416, minus actually moving the test_suite.h contents to a .cc file.
Cleanup in base. This moves the implementation (and a bunch of header file
dependencies) from the multiprocess test and the test_suite headers to .cc
files. Moves multiprocess_test to the test directory, and all of this stuff to
the existing base_test_support project. I also used the base namespace.

Previously other projects included this functionality just by #include because
it was all inline, so I had to add dependencies on base_test_support in a few
places.

Moves and renames the command line switch this was using to base_switches. Move
the base switch for process type to chrome switches.

TEST=none
BUG=none
Review URL: http://codereview.chromium.org/3026055


git-svn-id: http://src.chromium.org/svn/trunk/src/courgette@56399 4ff67af0-8c30-449e-8e8b-ad334ec8d88c