Factored out common ELF processing into a common superclass, an x86 subclass, and an initial ARM subclass.
Created a new disassembler virtual class, DisassemblerElf32, that
implements the common functions for processing ELF 32-bit files.  This
class is defined in disassembler_elf_32.{h|cc}, and has nearly all of
the methods and data that used to be in
disassembler_elf_32_x86.{h|cc}.  This class has two subclasses, one
for x86 and one for arm, which implement the architecture-specific
parts for generating courgette patches, namely extracting relative and
absolute addresses.  The ARM subclass is just a shell, but still
yields correct courgette patches.

The rest of the files add new enums for ARM and use the ARM subclasses
for the courgette command-line tool.

BUG=

Review URL: https://chromiumcodereview.appspot.com/17325003

git-svn-id: http://src.chromium.org/svn/trunk/src/courgette@207388 4ff67af0-8c30-449e-8e8b-ad334ec8d88c