Allow Win64 build of base library even in a Win32 build.
Some components of a 32 bit build of chrome for Windows need to be built
as Win64 Dlls.  To allow those components to use base, we make a Win64
build of the base library even when building chrome 32 bit.

This is needed by the Chrome Desk Band code. crbug.com/327435.
https://codereview.chromium.org/79173004/

BUG=327435

Review URL: https://codereview.chromium.org/103333004

git-svn-id: http://src.chromium.org/svn/trunk/src/courgette@243936 4ff67af0-8c30-449e-8e8b-ad334ec8d88c